+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.1 (2023-03-31)
 
+- Set type of 'duration' and 'timestamp' formats to convict.Format.
 - Fix repository and issues URLs in package.json.
 
 ## 1.0.0 (2023-02-18)